+++ src/features/extended-stack-error/extended-stack-error.constants.ts
@@ -1,0 +1,1 @@
+export const EXTENDED_STACK_ADDITIONAL_INFO_TEXT = 'Debug Info';