+++ src/features/extended-stack-error/extended-stack-error.utils.ts
@@ -1,0 +1,5 @@
+import * as Constants from './extended-stack-error.constants';
+
+export const getDebugStackInfo = (info: Record<string, any>): string => {
+  return ` \n\n${Constants.EXTENDED_STACK_ADDITIONAL_INFO_TEXT}: ${JSON.stringify(info, null, 2)}\n\n`;
+};